+++ packages/espressocash_api/lib/src/client.g.dart
@@ -23,7 +23,7 @@ class _EspressoCashClient implements EspressoCashClient {
   @override
   Future<CreateDirectPaymentResponseDto> createDirectPayment(
       CreateDirectPaymentRequestDto request) async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
@@ -51,10 +51,10 @@ class _EspressoCashClient implements EspressoCashClient {
 
   @override
   Future<GetFeesResponseDto> getFees() async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
-    final Map<String, dynamic>? _data = null;
+    const Map<String, dynamic>? _data = null;
     final _result = await _dio
         .fetch<Map<String, dynamic>>(_setStreamType<GetFeesResponseDto>(Options(
       method: 'POST',
@@ -78,7 +78,7 @@ class _EspressoCashClient implements EspressoCashClient {
 
   @override
   Future<SwapRouteResponseDto> getSwapRoute(SwapRouteRequestDto request) async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
@@ -107,7 +107,7 @@ class _EspressoCashClient implements EspressoCashClient {
   @override
   Future<CreatePaymentResponseDto> createPaymentEc(
       CreatePaymentRequestDto request) async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
@@ -136,7 +136,7 @@ class _EspressoCashClient implements EspressoCashClient {
   @override
   Future<ReceivePaymentResponseDto> receivePaymentEc(
       ReceivePaymentRequestDto request) async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
@@ -165,7 +165,7 @@ class _EspressoCashClient implements EspressoCashClient {
   @override
   Future<CancelPaymentResponseDto> cancelPaymentEc(
       CancelPaymentRequestDto request) async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
@@ -194,7 +194,7 @@ class _EspressoCashClient implements EspressoCashClient {
   @override
   Future<GenerateScalexLinkResponseDto> generateScalexLink(
       GenerateScalexLinkRequestDto request) async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
@@ -223,7 +223,7 @@ class _EspressoCashClient implements EspressoCashClient {
   @override
   Future<OrderStatusScalexResponseDto> fetchScalexTransaction(
       OrderStatusScalexRequestDto referenceId) async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
@@ -252,7 +252,7 @@ class _EspressoCashClient implements EspressoCashClient {
   @override
   Future<ScalexWithdrawResponseDto> createScalexWithdraw(
       ScalexWithdrawRequestDto request) async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
@@ -280,10 +280,10 @@ class _EspressoCashClient implements EspressoCashClient {
 
   @override
   Future<ScalexRateFeeResponseDto> fetchScalexFeesAndRate() async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
-    final Map<String, dynamic>? _data = null;
+    const Map<String, dynamic>? _data = null;
     final _result = await _dio.fetch<Map<String, dynamic>>(
         _setStreamType<ScalexRateFeeResponseDto>(Options(
       method: 'POST',
@@ -307,7 +307,7 @@ class _EspressoCashClient implements EspressoCashClient {
 
   @override
   Future<void> updateUserWalletCountry(WalletCountryRequestDto request) async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
@@ -333,7 +333,7 @@ class _EspressoCashClient implements EspressoCashClient {
   @override
   Future<PaymentQuoteResponseDto> getDlnQuote(
       PaymentQuoteRequestDto request) async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
@@ -362,7 +362,7 @@ class _EspressoCashClient implements EspressoCashClient {
   @override
   Future<OrderIdDlnResponseDto> fetchDlnOrderId(
       OrderIdDlnRequestDto request) async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
@@ -391,7 +391,7 @@ class _EspressoCashClient implements EspressoCashClient {
   @override
   Future<OrderStatusDlnResponseDto> fetchDlnStatus(
       OrderStatusDlnRequestDto request) async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
@@ -420,7 +420,7 @@ class _EspressoCashClient implements EspressoCashClient {
   @override
   Future<IncomingQuoteResponseDto> getIncomingDlnQuote(
       IncomingQuoteRequestDto request) async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
@@ -448,10 +448,10 @@ class _EspressoCashClient implements EspressoCashClient {
 
   @override
   Future<GetFreeNonceResponseDto> getFreeNonce() async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
-    final Map<String, dynamic>? _data = null;
+    const Map<String, dynamic>? _data = null;
     final _result = await _dio.fetch<Map<String, dynamic>>(
         _setStreamType<GetFreeNonceResponseDto>(Options(
       method: 'POST',
@@ -474,34 +474,65 @@ class _EspressoCashClient implements EspressoCashClient {
   }
 
   @override
-  Future<void> submitDurableTx(SubmitDurableTxRequestDto request) async {
-    const _extra = <String, dynamic>{};
+  Future<SubmitDurableTxResponseDto> submitDurableTx(
+      SubmitDurableTxRequestDto request) async {
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
     _data.addAll(request.toJson());
-    await _dio.fetch<void>(_setStreamType<void>(Options(
+    final _result = await _dio.fetch<Map<String, dynamic>>(
+        _setStreamType<SubmitDurableTxResponseDto>(Options(
       method: 'POST',
       headers: _headers,
       extra: _extra,
     )
-        .compose(
-          _dio.options,
-          '/submitDurableTx',
-          queryParameters: queryParameters,
-          data: _data,
-        )
-        .copyWith(
-            baseUrl: _combineBaseUrls(
-          _dio.options.baseUrl,
-          baseUrl,
-        ))));
+            .compose(
+              _dio.options,
+              '/submitDurableTx',
+              queryParameters: queryParameters,
+              data: _data,
+            )
+            .copyWith(
+                baseUrl: _combineBaseUrls(
+              _dio.options.baseUrl,
+              baseUrl,
+            ))));
+    final value = SubmitDurableTxResponseDto.fromJson(_result.data!);
+    return value;
+  }
+
+  @override
+  Future<GetDurableFeesResponseDto> getDurableFees() async {
+    final _extra = <String, dynamic>{};
+    final queryParameters = <String, dynamic>{};
+    final _headers = <String, dynamic>{};
+    const Map<String, dynamic>? _data = null;
+    final _result = await _dio.fetch<Map<String, dynamic>>(
+        _setStreamType<GetDurableFeesResponseDto>(Options(
+      method: 'POST',
+      headers: _headers,
+      extra: _extra,
+    )
+            .compose(
+              _dio.options,
+              '/getDurableFees',
+              queryParameters: queryParameters,
+              data: _data,
+            )
+            .copyWith(
+                baseUrl: _combineBaseUrls(
+              _dio.options.baseUrl,
+              baseUrl,
+            ))));
+    final value = GetDurableFeesResponseDto.fromJson(_result.data!);
+    return value;
   }
 
   @override
   Future<ShortenLinkResponseDto> shortenLink(
       ShortenLinkRequestDto request) async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
@@ -530,7 +561,7 @@ class _EspressoCashClient implements EspressoCashClient {
   @override
   Future<UnshortenLinkResponseDto> unshortenLink(
       UnshortenLinkRequestDto request) async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
@@ -558,7 +589,7 @@ class _EspressoCashClient implements EspressoCashClient {
 
   @override
   Future<GasFeeResponseDto> getGasFees(GasFeeRequestDto request) async {
-    const _extra = <String, dynamic>{};
+    final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
@@ -581,6 +612,62 @@ class _EspressoCashClient implements EspressoCashClient {
               baseUrl,
             ))));
     final value = GasFeeResponseDto.fromJson(_result.data!);
+    return value;
+  }
+
+  @override
+  Future<GetRatesResponseDto> getRates() async {
+    final _extra = <String, dynamic>{};
+    final queryParameters = <String, dynamic>{};
+    final _headers = <String, dynamic>{};
+    const Map<String, dynamic>? _data = null;
+    final _result = await _dio.fetch<Map<String, dynamic>>(
+        _setStreamType<GetRatesResponseDto>(Options(
+      method: 'POST',
+      headers: _headers,
+      extra: _extra,
+    )
+            .compose(
+              _dio.options,
+              '/rates',
+              queryParameters: queryParameters,
+              data: _data,
+            )
+            .copyWith(
+                baseUrl: _combineBaseUrls(
+              _dio.options.baseUrl,
+              baseUrl,
+            ))));
+    final value = GetRatesResponseDto.fromJson(_result.data!);
+    return value;
+  }
+
+  @override
+  Future<PriorityFeesResponseDto> getPriorityFeeEstimate(
+      PriorityFeesRequestDto request) async {
+    final _extra = <String, dynamic>{};
+    final queryParameters = <String, dynamic>{};
+    final _headers = <String, dynamic>{};
+    final _data = <String, dynamic>{};
+    _data.addAll(request.toJson());
+    final _result = await _dio.fetch<Map<String, dynamic>>(
+        _setStreamType<PriorityFeesResponseDto>(Options(
+      method: 'POST',
+      headers: _headers,
+      extra: _extra,
+    )
+            .compose(
+              _dio.options,
+              '/getPriorityFeeEstimate',
+              queryParameters: queryParameters,
+              data: _data,
+            )
+            .copyWith(
+                baseUrl: _combineBaseUrls(
+              _dio.options.baseUrl,
+              baseUrl,
+            ))));
+    final value = PriorityFeesResponseDto.fromJson(_result.data!);
     return value;
   }