+++ packages/espressocash_app/lib/features/activities/widgets/activity_tile.dart
@@ -76,6 +76,7 @@ class CpActivityTile extends StatelessWidget {
 const _titleStyle = TextStyle(
   fontSize: 16,
   letterSpacing: .23,
+  color: Colors.white,
   fontWeight: FontWeight.w600,
 );
 
@@ -88,6 +89,7 @@ const _inAmountStyle = TextStyle(
 
 const _subtitleStyle = TextStyle(
   fontSize: 14,
+  color: Colors.white,
   letterSpacing: .19,
 );