+++ packages/espressocash_app/lib/features/authenticated/screens/home_screen.dart
@@ -7,7 +7,6 @@ import '../../../gen/assets.gen.dart';
 import '../../../ui/loader.dart';
 import '../../../ui/navigation_bar/navigation_bar.dart';
 import '../../../ui/navigation_bar/navigation_button.dart';
-import '../../../utils/routing.dart';
 import '../../activities/screens/activities_screen.dart';
 import '../../dynamic_links/services/dynamic_links_notifier.dart';
 import '../../incoming_link_payments/widgets/pending_ilp_listener.dart';
@@ -23,8 +22,12 @@ class HomeScreen extends StatefulWidget {
   const HomeScreen({super.key});
 
   static void openWalletTab(BuildContext context) {
-    context.openFirstScreen();
-    context.read<TabNotifier>().value = 1;
+    final _HomeScreenState? state =
+        context.findAncestorStateOfType<_HomeScreenState>();
+    if (state != null) {
+      state._pageController.jumpToPage(1);
+      state._tabNotifier.value = 1;
+    }
   }
 
   static void openActivitiesTab(
@@ -32,14 +35,23 @@ class HomeScreen extends StatefulWidget {
     // ignore: avoid-unused-parameters, fix later
     ActivitiesTab tab = ActivitiesTab.pending,
   }) {
-    context.openFirstScreen();
-    context.read<TabNotifier>().value = 2;
+    final _HomeScreenState? state =
+        context.findAncestorStateOfType<_HomeScreenState>();
+    if (state != null) {
+      state._pageController.jumpToPage(3);
+      state._tabNotifier.value = 3;
+    }
   }
 
   static void openSwapTab(BuildContext context, {Token? initialToken}) {
     print(initialToken); //TODO pass to SwapScreen
-    context.openFirstScreen();
-    context.read<TabNotifier>().value = 3;
+
+    final _HomeScreenState? state =
+        context.findAncestorStateOfType<_HomeScreenState>();
+    if (state != null) {
+      state._pageController.jumpToPage(2);
+      state._tabNotifier.value = 2;
+    }
   }
 
   @override
@@ -48,10 +60,12 @@ class HomeScreen extends StatefulWidget {
 
 class _HomeScreenState extends State<HomeScreen> {
   final _tabNotifier = TabNotifier();
+  final _pageController = PageController();
 
   @override
   void dispose() {
     _tabNotifier.dispose();
+    _pageController.dispose();
     super.dispose();
   }
 
@@ -68,14 +82,22 @@ class _HomeScreenState extends State<HomeScreen> {
                     builder: (context, value, _) => Scaffold(
                       backgroundColor: Colors.white,
                       extendBody: true,
-                      body: _pages[value].builder(context),
+                      body: PageView(
+                        controller: _pageController,
+                        physics: const NeverScrollableScrollPhysics(),
+                        children:
+                            _pages.map((e) => e.builder(context)).toList(),
+                      ),
                       bottomNavigationBar: CPNavigationBar(
                         items: _pages
                             .mapIndexed(
                               (i, p) => CpNavigationButton(
                                 icon: p.icon,
                                 active: value == i,
-                                onPressed: () => _tabNotifier.value = i,
+                                onPressed: () {
+                                  _tabNotifier.value = i;
+                                  _pageController.jumpToPage(i);
+                                },
                               ),
                             )
                             .toList(),
@@ -114,9 +136,8 @@ class TabNotifier extends ValueNotifier<int> {
 }
 
 // ignore: avoid-function-type-in-records, fix later
-final List<({SvgGenImage icon, String path, WidgetBuilder builder})> _pages = [
+final List<({SvgGenImage icon, WidgetBuilder builder})> _pages = [
   (
-    path: '/home',
     icon: Assets.icons.home,
     builder: (context) => MainScreen(
           onSendMoneyPressed: () => HomeScreen.openWalletTab(context),
@@ -127,17 +148,14 @@ final List<({SvgGenImage icon, String path, WidgetBuilder builder})> _pages = [
         ),
   ),
   (
-    path: '/wallet',
     icon: Assets.icons.wallet,
     builder: (context) => const WalletScreen(),
   ),
   (
-    path: '/swap',
     icon: Assets.icons.swap,
     builder: (context) => const SwapScreen(),
   ),
   (
-    path: '/activities',
     icon: Assets.icons.notifications,
     builder: (context) => ActivitiesScreen(
           initialTab: ActivitiesTab.pending,