+++ packages/espressocash_app/lib/features/authenticated/screens/main_screen.dart
@@ -1,5 +1,6 @@
 import 'dart:math';
 
+import 'package:fast_immutable_collections/fast_immutable_collections.dart';
 import 'package:flutter/material.dart';
 
 import '../../../di.dart';
@@ -9,11 +10,14 @@ import '../../../ui/theme.dart';
 import '../../../ui/value_stream_builder.dart';
 import '../../activities/services/tx_updater.dart';
 import '../../activities/widgets/recent_activity.dart';
-import '../../conversion_rates/services/watch_user_fiat_balance.dart';
+import '../../balances/data/repository.dart';
+import '../../currency/models/amount.dart';
+import '../../tokens/token.dart';
 import '../widgets/home_add_cash.dart';
 import '../widgets/home_app_bar.dart';
 import '../widgets/home_carousel.dart';
 import '../widgets/investment_header.dart';
+import '../widgets/portfolio_widget.dart';
 import '../widgets/refresh_balance_wrapper.dart';
 
 class MainScreen extends StatelessWidget {
@@ -29,9 +33,13 @@ class MainScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) => CpTheme.dark(
         child: ValueStreamBuilder<bool>(
-          create: () =>
-              sl<WatchUserFiatBalance>().call().map((it) => it.isZero),
-          builder: (context, isZeroAmount) => isZeroAmount
+          create: () => (
+            sl<TokenBalancesRepository>()
+                .watchUserTokens()
+                .map((it) => it.isEmpty),
+            true
+          ),
+          builder: (context, isEmpty) => isEmpty
               ? const HomeAddCashContent()
               : _MainContent(
                   onSendMoneyPressed: onSendMoneyPressed,
@@ -84,9 +92,22 @@ class _MainContent extends StatelessWidget {
                   ),
                 ),
                 SliverToBoxAdapter(
-                  child: HomeCarouselWidget(
-                    onSendMoneyPressed: onSendMoneyPressed,
+                  child: ValueStreamBuilder<IList<CryptoAmount>>(
+                    create: () => (
+                      sl<TokenBalancesRepository>().watchTokenBalances(
+                        ignoreTokens: [Token.usdc],
+                      ),
+                      const IListConst([])
+                    ),
+                    builder: (context, tokens) => tokens.isNotEmpty
+                        ? const SizedBox.shrink()
+                        : HomeCarouselWidget(
+                            onSendMoneyPressed: onSendMoneyPressed,
+                          ),
                   ),
+                ),
+                const SliverToBoxAdapter(
+                  child: PortfolioWidget(),
                 ),
                 SliverToBoxAdapter(
                   child: RecentActivityWidget(