+++ packages/espressocash_app/lib/features/balances/services/balances_bloc.dart
@@ -1,7 +1,8 @@
 import 'dart:async';
 
 import 'package:bloc_concurrency/bloc_concurrency.dart';
-
+import 'package:collection/collection.dart';
+import 'package:dfunc/dfunc.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:injectable/injectable.dart';
@@ -16,7 +17,8 @@ import '../../analytics/analytics_manager.dart';
 import '../../currency/models/amount.dart';
 import '../../currency/models/currency.dart';
 import '../../tokens/token.dart';
-import '../data/balances_repository.dart';
+import '../../tokens/token_list.dart';
+import '../data/repository.dart';
 
 part 'balances_bloc.freezed.dart';
 
@@ -28,14 +30,16 @@ class BalancesBloc extends Bloc<BalancesEvent, BalancesState>
     with DisposableBloc {
   BalancesBloc(
     this._solanaClient,
-    this._repository,
+    this._tokens,
+    this._tokensRepository,
     this._analyticsManager,
   ) : super(const ProcessingStateNone()) {
     on<BalancesEventRequested>(_handleRequested, transformer: droppable());
   }
 
   final SolanaClient _solanaClient;
-  final BalancesRepository _repository;
+  final TokenList _tokens;
+  final TokenBalancesRepository _tokensRepository;
   final AnalyticsManager _analyticsManager;
 
   Future<void> _handleRequested(
@@ -45,18 +49,50 @@ class BalancesBloc extends Bloc<BalancesEvent, BalancesState>
     try {
       emit(const ProcessingState.processing());
 
-      final usdcBalance = await _solanaClient.getUsdcBalance(event.address);
+      final sol = await _solanaClient.getSolBalance(event.address);
+
+      final allAccounts = await _solanaClient.getSplAccounts(event.address);
+      final mainAccounts = await Future.wait<_MainTokenAccount?>(
+        allAccounts.map((programAccount) async {
+          final account = programAccount.account;
+          final data = account.data;
+
+          if (data is ParsedAccountData) {
+            return data.maybeWhen<Future<_MainTokenAccount?>>(
+              splToken: (parsed) => parsed.maybeMap<Future<_MainTokenAccount?>>(
+                account: (a) => _MainTokenAccount.create(
+                  programAccount.pubkey,
+                  a.info,
+                  _tokens,
+                ),
+                orElse: () async => null,
+              ),
+              orElse: () async => null,
+            );
+          }
+        }),
+      ).then(compact);
+
+      final tokenBalances = mainAccounts.map(
+        (a) => CryptoAmount(
+          value: int.parse(a.info.tokenAmount.amount),
+          cryptoCurrency: CryptoCurrency(token: a.token),
+        ),
+      );
+
+      final usdcBalance = tokenBalances.firstWhereOrNull(
+        (balance) => balance.cryptoCurrency.token == Token.usdc,
+      );
 
       if (isClosed) return;
 
       emit(const ProcessingState.none());
 
-      if (usdcBalance == null) {
-        return;
+      if (usdcBalance != null) {
+        _analyticsManager.setUsdcBalance(usdcBalance.decimal);
       }
 
-      _analyticsManager.setUsdcBalance(usdcBalance.decimal);
-      _repository.save(usdcBalance);
+      await _tokensRepository.save([...tokenBalances, sol]);
     } on Exception catch (exception) {
       _logger.severe('Failed to fetch balances', exception);
 
@@ -66,6 +102,32 @@ class BalancesBloc extends Bloc<BalancesEvent, BalancesState>
       emit(const ProcessingState.none());
     }
   }
+}
+
+class _MainTokenAccount {
+  const _MainTokenAccount._(this.pubKey, this.info, this.token);
+
+  static Future<_MainTokenAccount?> create(
+    String pubKey,
+    SplTokenAccountDataInfo info,
+    TokenList tokens,
+  ) async {
+    final expectedPubKey = await findAssociatedTokenAddress(
+      owner: Ed25519HDPublicKey.fromBase58(info.owner),
+      mint: Ed25519HDPublicKey.fromBase58(info.mint),
+    );
+
+    if (expectedPubKey.toBase58() != pubKey) return null;
+
+    final token = tokens.findTokenByMint(info.mint);
+
+    // TODO(IA): we should find a way to display this
+    return token == null ? null : _MainTokenAccount._(pubKey, info, token);
+  }
+
+  final Token token;
+  final SplTokenAccountDataInfo info;
+  final String pubKey;
 }
 
 class BalancesRequestException implements Exception {
@@ -80,26 +142,23 @@ sealed class BalancesEvent with _$BalancesEvent {
 }
 
 extension on SolanaClient {
-  Future<CryptoAmount?> getUsdcBalance(String address) async {
-    try {
-      final usdcTokenAccount = await findAssociatedTokenAddress(
-        owner: Ed25519HDPublicKey.fromBase58(address),
-        mint: Ed25519HDPublicKey.fromBase58(Token.usdc.address),
-      );
+  Future<CryptoAmount> getSolBalance(String address) async {
+    final lamports = await rpcClient
+        .getBalance(
+          address,
+          commitment: Commitment.confirmed,
+        )
+        .value;
 
-      final balance = await rpcClient
-          .getTokenAccountBalance(
-            usdcTokenAccount.toBase58(),
-            commitment: Commitment.confirmed,
-          )
-          .value;
-
-      return CryptoAmount(
-        value: int.parse(balance.amount),
-        cryptoCurrency: Currency.usdc,
-      );
-    } on Exception {
-      return null;
-    }
+    return CryptoAmount(value: lamports, cryptoCurrency: Currency.sol);
   }
+
+  Future<Iterable<ProgramAccount>> getSplAccounts(String address) => rpcClient
+      .getTokenAccountsByOwner(
+        address,
+        const TokenAccountsFilter.byProgramId(TokenProgram.programId),
+        commitment: Commitment.confirmed,
+        encoding: Encoding.jsonParsed,
+      )
+      .value;
 }