+++ packages/espressocash_app/lib/features/tokens/token_list.dart
@@ -1,6 +1,4 @@
 import 'package:collection/collection.dart';
-import 'package:dfunc/dfunc.dart';
-
 import 'package:json_annotation/json_annotation.dart';
 import 'package:meta/meta.dart' show visibleForTesting;
 
@@ -53,33 +51,10 @@ class TokenList {
 
   // ignore: avoid-non-null-assertion, required here
   Token requireTokenByMint(String mint) => findTokenByMint(mint)!;
-
-  Token fromCoingecko({
-    required String? coingeckoId,
-    required String? symbol,
-    required String? name,
-    required String? image,
-  }) =>
-      tokens
-          .singleWhereOrNull(
-            (t) =>
-                (symbol?.toLowerCase().contains(t.symbol.toLowerCase()) ??
-                    false) &&
-                t.coingeckoId?.toLowerCase() == coingeckoId?.toLowerCase(),
-          )
-          .ifNull(
-            () => _createStubToken(
-              coingeckoId: coingeckoId,
-              symbol: symbol,
-              name: name,
-              image: image,
-            ),
-          );
 }
 
 extension TokenExt on Iterable<Token> {
-  Iterable<String> get coingeckoIds =>
-      this.map((t) => t.coingeckoId).whereType<String>();
+  Iterable<String> get symbols => map((t) => t.symbol);
 
   Iterable<Token> filterTokens(String searchTerm) => where(
         (token) =>
@@ -88,20 +63,3 @@ extension TokenExt on Iterable<Token> {
             token.name.toLowerCase().contains(searchTerm.toLowerCase()),
       );
 }
-
-Token _createStubToken({
-  required String? coingeckoId,
-  required String? symbol,
-  required String? name,
-  required String? image,
-}) =>
-    Token(
-      chainId: currentChainId,
-      address: coingeckoId ?? '',
-      symbol: symbol?.toUpperCase() ?? '',
-      name: name ?? '',
-      decimals: 0,
-      logoURI: image,
-      tags: const [],
-      extensions: Extensions(coingeckoId: coingeckoId),
-    );