+++ packages/espressocash_app/CHANGELOG.md
@@ -1,3 +1,32 @@
+## 1.121.1
+
+ - **FIX**: do not wait for intercom login (#1493).
+ - **FIX**: profile spacing colors (#1490).
+ - **FIX**: sol conversion matching (#1491).
+
+## 1.121.0
+
+ - **REFACTOR**: balances module (#1476).
+ - **FIX**: main screen flow (#1487).
+ - **FIX**: ignore usdc token on crypto portfolio (#1478).
+ - **FIX**: activity tile text color to white (#1477).
+ - **FEAT**: log total crypto balance (#1488).
+ - **FEAT**: sort crypto portfolio list (#1480).
+ - **FEAT**: home & payment improved buttons (#1468).
+
+## 1.120.0
+
+ - **REFACTOR**: request feature flag cleanup (#1455).
+ - **REFACTOR**: remove tx failure logging (#1453).
+ - **REFACTOR**: enable profiling (#1447).
+ - **FIX**: transaction request fixes (#1449).
+ - **FEAT**: show users list of tokens (#1433).
+ - **FEAT**: black background in activity and profile page (#1461).
+ - **FEAT**: fetch usd rate from backend (#1465).
+ - **FEAT**: homepage theme (#1452).
+ - **FEAT**: update local token list data (#1454).
+ - **FEAT**: update analytics (#1451).
+
 ## 1.119.0
 
  - **FEAT**: optmize startup time (#1444).